Adjust the pan table to produce constant power output
The original pan-table calculation per the InterWave(TM) IC
Am78C201/202 Programmer's Guide, Rev.2, 1996 does not produce
constant power as stated on page 164 of the Gravis UltraSound
Software SDK rev 2.33: "As the balance is shifted from left to right,
the total output power of both channels is constant.".

To correct this, we adjust the original table values per the equation
for constant power: (sqrt(2)/2)*(cos(angle) +/- sin(angle).
Reference: Roads, Curtis (1996). The Computer Music Tutorial.
Cambridge: The MIT Press, pp. 457–461.

The results is roughly 2.5x less power given to the faded side, and
more power given to the side in which the software has panned toward.
As 3dB is twice the perceived volume, 2.5x exceeds 3dB and thus is a
noticeable adjustment.